requirements.txt
# Requirements - copy of install_requires in setup.py
six>=1.9.0
requests>=2.7.0

# Dev requirements
pillow>=2.8.2

# Test requirements
pytest
coverage
mock
